// src/app/api/mmli-backend/v1/model/bodyCreateJobJobTypeJobsPost.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BodyCreateJobJobTypeJobsPost { 
    job_id?: string;
    run_id?: string;
    email?: string;
    job_info?: string;
}