// src/app/api/mmli-backend/v1/model/checkReactionSiteRequest.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CheckReactionSiteRequest { 
    input: string;
    role: CheckReactionSiteRequest.RoleEnum;
    input_type: CheckReactionSiteRequest.InputTypeEnum;
}
export namespace CheckReactionSiteRequest {
    export type RoleEnum = 'el' | 'nuc';
    export const RoleEnum = {
        El: 'el' as RoleEnum,
        Nuc: 'nuc' as RoleEnum
    };
    export type InputTypeEnum = 'smi' | 'cml' | 'cdxml';
    export const InputTypeEnum = {
        Smi: 'smi' as InputTypeEnum,
        Cml: 'cml' as InputTypeEnum,
        Cdxml: 'cdxml' as InputTypeEnum
    };
}